Remove Google Mobile Ads SDK; use no-op ads provider.
