+++ Package.swift
@@ -15,21 +15,16 @@ let package = Package(
     ],
     dependencies: [
         .package(
-            url: "https://github.com/googleads/swift-package-manager-google-mobile-ads.git",
-            exact: "13.1.0"
-        ),
-        .package(
             url: "https://github.com/BastiaanJansen/Toast-Swift.git",
             exact: "2.1.3"
         ),
-            ],
+    ],
     targets: [
         .target(
             name: "SwapFoundationKit",
             dependencies: [
-                .product(name: "GoogleMobileAds", package: "swift-package-manager-google-mobile-ads"),
                 .product(name: "Toast", package: "Toast-Swift"),
-                            ]
+            ]
         ),
         .testTarget(
             name: "SwapFoundationKitTests",
