+++ Sources/SwapFoundationKit/Ads/AdsProvider.swift
@@ -9,7 +9,6 @@
  * Refer to the COPYING file of the official project for license.
  *****************************************************************************/
 
-#if !targetEnvironment(simulator) && canImport(UIKit) && canImport(GoogleMobileAds)
 import UIKit
 
 @MainActor
@@ -38,12 +37,38 @@ protocol AdsProviderFactory {
 }
 
 @MainActor
+private final class NoOpAdsProvider: AdsProvider {
+    func start() async { }
+
+    func preload(
+        _ placement: AdPlacement,
+        adUnitID: String,
+        eventHandler: @escaping @MainActor (AdLifecycleEvent) -> Void
+    ) { }
+
+    func present(
+        _ placement: AdPlacement,
+        adUnitID: String,
+        from viewController: UIViewController,
+        eventHandler: @escaping @MainActor (AdLifecycleEvent) -> Void
+    ) async -> AdPresentationResult {
+        .unavailable
+    }
+
+    func makeBannerViewController(
+        adUnitID: String,
+        eventHandler: @escaping @MainActor (AdLifecycleEvent) -> Void
+    ) -> UIViewController {
+        EmptyBannerViewController()
+    }
+}
+
+@MainActor
 struct DefaultAdsProviderFactory: AdsProviderFactory {
     func makeProvider(from configuration: AdsProviderConfiguration) -> AdsProvider {
         switch configuration {
-        case .google(let googleConfiguration):
-            return GoogleAdsProvider(configuration: googleConfiguration)
+        case .google:
+            return NoOpAdsProvider()
         }
     }
 }
-#endif
